chore: introduce the project's `biome.json` configuration
This commit introduces the configuration file of the
[BiomeJS](https://biomejs.dev) tool for formatting and linting the
application codebase. The main rules applied are the recommended linting
warning/errors, the JavaScript formatting of semi-colons as needed,
always wrapping parenthesis, triailing comma present in all
circumstances, single quotes for default JavaScript code and double
quotes for JSX expression attributes/props. In addition, this
configuration also establishes the indent with space and the line width
of 140.