feat: introduce the `Toast` composite component
This commit introduces the `Toast`, `ToastAction`, `ToastClose`,
`ToastContent`, `ToastController`, `ToastDescription`, `ToastProvider`,
and `ToastTitle` components. Also, this commit defines the `toast()`
util function and the `useToast()` custom React hook to call and manage
the appearance and status of the toast on the demand without requiring
defining it inside a JSX expression or component.

This component is heavily based on the [shadcn/ui](https://ui.shadcn.com)
and `sonner` component.